Cleanup and linter fixes
