+++ lib/class-converter.php
@@ -7,8 +7,8 @@
 
 namespace NewspackContentConverter;
 
+use \NewspackContentConverter\ConverterController;
 use \NewspackContentConverter\Installer;
-use \NewspackContentConverter\ConversionProcessor;
 
 /**
  * Content Converter.
@@ -23,26 +23,19 @@ class Converter {
 	private $installer;
 
 	/**
-	 * The conversion processor service.
-	 *
-	 * @var ConversionProcessor
-	 */
-	private $conversion_processor;
-
-	/**
 	 * Converter constructor.
 	 *
 	 * @param Installer           $installer The installer service.
-	 * @param ConversionProcessor $conversion_processor The conversion processor service.
+	 * @param ConverterController $controller The main Controller.
 	 */
-	public function __construct( Installer $installer, ConversionProcessor $conversion_processor ) {
-		$this->installer            = $installer;
-		$this->conversion_processor = $conversion_processor;
+	public function __construct( Installer $installer, ConverterController $controller ) {
+		$this->installer  = $installer;
+		$this->controller = $controller;
 
 		$this->register_installation_hook();
 		$this->add_admin_menu();
-		$this->enqueue_newspack_block_editor_assets();
-		$this->register_api_routes();
+		add_action( 'admin_enqueue_scripts', [ $this, 'enqueue_scripts' ] );
+		add_action( 'rest_api_init', [ $this->controller, 'register_routes' ] );
 		$this->disable_autosave_posts();
 	}
 
@@ -129,13 +122,6 @@ class Converter {
 	}
 
 	/**
-	 * Enqueues assets.
-	 */
-	private function enqueue_newspack_block_editor_assets() {
-		add_action( 'admin_enqueue_scripts', [ $this, 'enqueue_scripts' ] );
-	}
-
-	/**
 	 * Enqueues script assets.
 	 */
 	public function enqueue_scripts() {
@@ -205,364 +191,5 @@ class Converter {
 			'post-new.php' === $pagenow &&
 			isset( $_GET['newspack-content-converter'] )
 		);
-	}
-
-	/**
-	 * Registers routes.
-	 */
-	private function register_api_routes() {
-
-		// Fetches info for the settings page.
-		add_action(
-			'rest_api_init',
-			function () {
-				register_rest_route(
-					'newspack-content-converter',
-					'/settings/get-info',
-					[
-						'methods'             => 'GET',
-						'callback'            => [ $this, 'api_settings_get_info' ],
-						'permission_callback' => [ $this, 'newspack_content_converter_rest_permission' ],
-					]
-				);
-			}
-		);
-
-		// Fetches info for the conversion page.
-		add_action(
-			'rest_api_init',
-			function () {
-				register_rest_route(
-					'newspack-content-converter',
-					'/conversion/get-info',
-					[
-						'methods'             => 'GET',
-						'callback'            => [ $this, 'api_conversion_get_info' ],
-						'permission_callback' => [ $this, 'newspack_content_converter_rest_permission' ],
-					]
-				);
-			}
-		);
-
-		// Initializes the conversion queue.
-		add_action(
-			'rest_api_init',
-			function () {
-				register_rest_route(
-					'newspack-content-converter',
-					'/conversion/initialize',
-					[
-						'methods'             => 'POST',
-						'callback'            => [ $this, 'api_conversion_initialize' ],
-						'permission_callback' => [ $this, 'newspack_content_converter_rest_permission' ],
-					]
-				);
-			}
-		);
-
-		// Fetches a batch to be converted to blocks.
-		add_action(
-			'rest_api_init',
-			function () {
-				register_rest_route(
-					'newspack-content-converter',
-					'/conversion/get-batch-data',
-					[
-						'methods'             => 'GET',
-						'callback'            => [ $this, 'api_conversion_get_batch_data' ],
-						'permission_callback' => [ $this, 'newspack_content_converter_rest_permission' ],
-					]
-				);
-			}
-		);
-
-		// Fetches info for the patching page.
-		add_action(
-			'rest_api_init',
-			function () {
-				register_rest_route(
-					'newspack-content-converter',
-					'/patching/get-info',
-					[
-						'methods'             => 'GET',
-						'callback'            => [ $this, 'api_patching_get_info' ],
-						'permission_callback' => [ $this, 'newspack_content_converter_rest_permission' ],
-					]
-				);
-			}
-		);
-
-		// Initializes patching.
-		add_action(
-			'rest_api_init',
-			function () {
-				register_rest_route(
-					'newspack-content-converter',
-					'/patching/initialize',
-					[
-						'methods'             => 'POST',
-						'callback'            => [ $this, 'api_patching_initialize' ],
-						'permission_callback' => [ $this, 'newspack_content_converter_rest_permission' ],
-					]
-				);
-			}
-		);
-
-		// Processes the next patching batch.
-		add_action(
-			'rest_api_init',
-			function () {
-				register_rest_route(
-					'newspack-content-converter',
-					'/patching/process-next-batch',
-					[
-						'methods'             => 'GET',
-						'callback'            => [ $this, 'api_patching_process_next_batch' ],
-						'permission_callback' => [ $this, 'newspack_content_converter_rest_permission' ],
-					]
-				);
-			}
-		);
-
-		// Fetches post_content.
-		add_action(
-			'rest_api_init',
-			function () {
-				register_rest_route(
-					'newspack-content-converter',
-					'/get-post-content-by-id',
-					[
-						'methods'             => 'POST',
-						'callback'            => [ $this, 'api_get_post_content_by_id' ],
-						'permission_callback' => [ $this, 'newspack_content_converter_rest_permission' ],
-					]
-				);
-			}
-		);
-
-		// Updates the converted Post content.
-		add_action(
-			'rest_api_init',
-			function () {
-				register_rest_route(
-					'newspack-content-converter',
-					'/conversion/update-post',
-					[
-						'methods'             => 'POST',
-						'callback'            => [ $this, 'api_update_converted_post_content' ],
-						'permission_callback' => [ $this, 'newspack_content_converter_rest_permission' ],
-					]
-				);
-			}
-		);
-	}
-
-	/**
-	 * Permission check, common to basic Converter endpoints.
-	 *
-	 * @return bool|WP_Error
-	 */
-	public function newspack_content_converter_rest_permission() {
-		$is_user_authorized = current_user_can( 'edit_posts' );
-
-		if ( ! $is_user_authorized ) {
-			return new WP_Error( 'newspack_content_converter_rest_invalid_permission', __( 'Unauthorized access.' ) );
-		}
-
-		return true;
-	}
-
-	/**
-	 * Callback for the /settings/get-info route.
-	 * Fetches info for the settings page.
-	 *
-	 * @return array Info for the settings page.
-	 */
-	public function api_settings_get_info() {
-		$content_types_csv     = $this->conversion_processor->get_conversion_content_types();
-		$content_statuses_csv  = $this->conversion_processor->get_conversion_content_statuses();
-		$conversion_batch_size = $this->conversion_processor->get_conversion_batch_size();
-		$patching_batch_size   = $this->conversion_processor->get_patching_batch_size();
-		$queued_entries        = $this->conversion_processor->get_queued_entries_total_number();
-
-		return [
-			'info' => [
-				'conversionContentTypesCsv'    => $content_types_csv,
-				'conversionContentStatusesCsv' => $content_statuses_csv,
-				'conversionBatchSize'          => $conversion_batch_size,
-				'patchingBatchSize'            => $patching_batch_size,
-				'queuedEntries'                => $queued_entries,
-			],
-		];
-	}
-
-	/**
-	 * Callback for the /conversion/get-info route.
-	 * Fetches info for the conversion page.
-	 *
-	 * @return array Info for the settings page.
-	 */
-	public function api_conversion_get_info() {
-		$is_conversion_ongoing = $this->conversion_processor->is_conversion_queued();
-		$queued_entries        = $this->conversion_processor->get_queued_entries_total_number();
-		$conversion_batch_size = $this->conversion_processor->get_conversion_batch_size();
-		$queued_batches        = $this->conversion_processor->get_conversion_queued_batches();
-		$max_batch             = $this->conversion_processor->get_conversion_max_batch();
-
-		return [
-			'info' => [
-				'isConversionOngoing' => $is_conversion_ongoing ? '1' : '0',
-				'queuedEntries'       => $queued_entries,
-				'conversionBatchSize' => $conversion_batch_size,
-				'queuedBatchesCsv'    => implode( ',', $queued_batches ),
-				'maxBatch'            => $max_batch,
-			],
-		];
-	}
-
-	/**
-	 * Callback for the /conversion/initialize route.
-	 * Initializes the conversion queue.
-	 *
-	 * @param WP_REST_Request $params JSON param, key 'request' with value 'initialize'.
-	 * @return array Formatted response.
-	 */
-	public function api_conversion_initialize( $params ) {
-		$json_params   = $params->get_json_params();
-		$request_param = isset( $json_params['request'] ) ? $json_params['request'] : null;
-
-		if ( ! $request_param || 'initialize' !== $request_param ) {
-			return;
-		}
-
-		$initialized = $this->conversion_processor->initialize_conversion();
-
-		return ( true === $initialized ) ? [ 'result' => 'queued' ] : null;
-	}
-
-	/**
-	 * Callback for the /conversion/get-batch-data route.
-	 * Fetches a batch to be converted to blocks.
-	 *
-	 * @return array Conversion batch data.
-	 */
-	public function api_conversion_get_batch_data() {
-		$ids        = $this->conversion_processor->set_next_conversion_batch_to_queue();
-		$this_batch = max( $this->conversion_processor->get_conversion_queued_batches() );
-		$max_batch  = $this->conversion_processor->get_conversion_max_batch();
-
-		return [
-			'ids'       => $ids,
-			'thisBatch' => $this_batch,
-			'maxBatch'  => $max_batch,
-		];
-	}
-
-	/**
-	 * Callback for the /patching/get-info route.
-	 * Fetches info for the patching page.
-	 *
-	 * @return array Info for the patching page.
-	 */
-	public function api_patching_get_info() {
-		$is_patching_ongoing     = $this->conversion_processor->is_patching_queued();
-		$queued_batches_patching = $this->conversion_processor->get_patching_queued_batches();
-		$max_batch_patching      = $this->conversion_processor->get_patching_max_batch();
-		$patching_batch_size     = $this->conversion_processor->get_patching_batch_size();
-		$queued_entries          = $this->conversion_processor->get_queued_entries_total_number();
-
-		return [
-			'info' => [
-				'isPatchingOngoing'        => $is_patching_ongoing,
-				'queuedBatchesPatchingCsv' => implode( ',', $queued_batches_patching ),
-				'maxBatchPatching'         => $max_batch_patching,
-				'patchingBatchSize'        => $patching_batch_size,
-				'queuedEntries'            => $queued_entries,
-			],
-		];
-
-	}
-
-	/**
-	 * Callback for the /patching/initialize route.
-	 * Initializes patching.
-	 *
-	 * @param WP_REST_Request $params JSON param, key 'request' with value 'initialize'.
-	 * @return array Formatted response.
-	 */
-	public function api_patching_initialize( $params ) {
-		$json_params   = $params->get_json_params();
-		$request_param = isset( $json_params['request'] ) ? $json_params['request'] : null;
-
-		if ( ! $request_param || 'initialize' !== $request_param ) {
-			return;
-		}
-
-		$initialized = $this->conversion_processor->initialize_patching();
-
-		return ( true === $initialized ) ? [ 'result' => 'queued' ] : null;
-	}
-
-	/**
-	 * Callback for the /patching/process-next-batch route.
-	 * /patching/process-next-batch
-	 *
-	 * @return array Formatted response.
-	 */
-	public function api_patching_process_next_batch() {
-		if ( ! $this->conversion_processor->is_patching_queued() ) {
-			return;
-		}
-
-		$current_batch = $this->conversion_processor->move_next_patching_batch_to_queue();
-		if ( false === $current_batch ) {
-			return;
-		}
-
-		$patched = $this->conversion_processor->apply_patches_to_batch( $current_batch );
-		if ( false == $patched ) {
-			return;
-		}
-
-		return [
-			'result' => 'patched',
-		];
-	}
-
-	/**
-	 * Callback for the /get-post-content-by-id route.
-	 * Fetches post_content.
-	 *
-	 * @param WP_REST_Request $params Params: 'id' Post ID.
-	 * @return array Post content.
-	 */
-	public function api_get_post_content_by_id( $params ) {
-		$json_params = $params->get_json_params();
-		$post_id     = isset( $json_params['id'] ) ? $json_params['id'] : null;
-
-		if ( ! $post_id ) {
-			return;
-		}
-
-		return $this->conversion_processor->get_post_content_by_id( $post_id );
-	}
-
-	/**
-	 * Callable for /conversion/update-post API endpoint.
-	 * Updates the converted Post content.
-	 *
-	 * @param WP_REST_Request $params Params: 'id' Post ID, 'content' Post content.
-	 */
-	public function api_update_converted_post_content( $params ) {
-		$json_params    = $params->get_json_params();
-		$post_id        = isset( $json_params['post_id'] ) ? $json_params['post_id'] : null;
-		$content_html   = isset( $json_params['content_html'] ) ? $json_params['content_html'] : null;
-		$content_blocks = isset( $json_params['content_blocks'] ) ? $json_params['content_blocks'] : null;
-
-		if ( ! $post_id || ! $content_html || ! $content_blocks ) {
-			return;
-		}
-
-		$this->conversion_processor->save_converted_post_content( $post_id, $content_html, $content_blocks );
 	}
 }
